requirements.txt
#python3.11.8
chromadb==0.4.24
datetime
langchain==0.1.13
matplotlib==3.8.3
ollama==0.1.7
pandas==2.2.1
pillow==10.2.0
psutil==5.9.8
pypdf==4.1.0
python-docx==1.1.0
sentence-transformers==2.6.0
streamlit==1.32.2
unstructured==0.10.8

